requirements.txt
adjustText==1.3.0
aiohttp==3.11.18
aiosqlite==0.21.0
beautifulsoup4==4.13.4
discord.py==2.4.0
Faker==37.1.0
fuzzywuzzy==0.18.0
httpx==0.28.1
matplotlib==3.10.1
numpy==2.2.5
openai==1.77.0
pandas==2.2.3
Pillow==11.2.1
python-dotenv==1.1.0
python_dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
Requests==2.32.3
scipy==1.15.2
